plumbing tile buffer through to config. WIP. #107
