+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.6.0 (2018-)
+- Fixed: 
+
 ## 0.5.0 (2017-12-12)
 
 - Added: Command line `cache seed` and `cache purge` commands (#64)
